Redesign global settings modal
